+++ mqt/MqtBase/BandList.h
@@ -57,8 +57,8 @@ class BandList
         QVector<BandInfo> bandList;
         bool parseFile ( const QString &bandFile );
         bool findBand ( const QString &freq, BandInfo & );
-        bool findBand ( int freq, BandInfo & );
-        bool findBand(double freq, BandInfo &bi);
+        bool findBand ( long freq, BandInfo & );
+        bool findBand ( double freq, BandInfo &bi);
 
         static BandList &getBandList();
